sandbox: add union file system to merge OCI layers

+++ internal/sandbox/syscall_linux.go
@@ -10,7 +10,7 @@ const sizeOfDirent = 19
 
 type dirent struct {
 	ino    uint64
-	off    int64
+	off    uint64
 	reclen uint16
 	typ    uint8
 }
@@ -18,7 +18,7 @@ type dirent struct {
 func makeDirent(typ uint8, ino, off uint64, name string) dirent {
 	return dirent{
 		ino:    ino,
-		off:    int64(off),
+		off:    off,
 		reclen: sizeOfDirent + uint16(len(name)) + 1,
 		typ:    typ,
 	}
